feat: bundle trusty-console into all daemons + mpm via Single-Install Convention (refs #943)
- Refactor trusty-console into a lib + bin crate: add [lib] (rlib, src/lib.rs),
  expose `pub async fn run()`, make all submodules `pub mod`, replace main.rs
  with a 17-line shim. Mirrors the trusty-embedderd (#187) and trusty-bm25-daemon
  (#190) patterns exactly. Remove `publish = false`; replace `exclude` list with
  `include` list for clean crates.io tarball.
- Fix clippy: add `Default` impl for `PollerCache` (new_without_default lint).
- Root Cargo.toml: add `trusty-console` to [workspace.dependencies] and
  [patch.crates-io] alongside trusty-embedderd / trusty-bm25-daemon.
- Add bundled [[bin]] shim + dep to 5 host crates:
  - trusty-search (src/bin/trusty-console.rs) → 0.24.1 → 0.24.2
  - trusty-memory (src/bin/trusty_console.rs) → 0.15.0 → 0.15.1
  - trusty-analyze (src/bin/trusty_console.rs, new src/bin/) → 0.5.1 → 0.5.2
  - trusty-review  (src/bin/trusty_console.rs, new src/bin/) → 0.3.6 → 0.3.7
  - trusty-mpm (src/bin/trusty_console.rs, publish=false — no version bump)
- Install proof: `cargo install --path crates/trusty-review --root /tmp/test`
  produces both `trusty-review` AND `trusty-console`.
