+++ crates/trusty-search/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "trusty-search"
-version = "0.24.1"
+version = "0.24.2"
 edition = "2021"
 description = "Machine-wide hybrid code search service: BM25 + vector + KG, zero cold-start, MCP server"
 readme = "README.md"
@@ -41,6 +41,14 @@ path = "src/bin/trusty-embedderd.rs"
 # single command (issue #187 Phase 2 follow-up). The daemon's logic lives
 # entirely in the trusty-embedderd library crate; this is a thin shim.
 
+[[bin]]
+name = "trusty-console"
+path = "src/bin/trusty-console.rs"
+# Why: bundle trusty-console into trusty-search's install surface so
+# `cargo install trusty-search` produces trusty-console alongside the other
+# binaries (Single-Install Convention, epic #943). All dashboard logic lives
+# in the trusty-console library crate; this is a thin shim.
+
 # [patch.crates-io] moved to workspace root Cargo.toml — Cargo only honours
 # [patch] tables at the workspace root. The monorepo root patches
 # trusty-common / trusty-embedder / trusty-mcp-core / trusty-symgraph to the
@@ -50,7 +58,7 @@ path = "src/bin/trusty-embedderd.rs"
 # ── Shared trusty-common crates ────────────────────────────────────────────
 trusty-common = { version = "0.14.1", default-features = false, features = ["axum-server", "mcp", "embedder", "embedder-test-support", "symgraph", "monitor-tui", "embedder-client", "bm25", "migrations", "cli-help", "update-check", "bug-capture"] }  # `symgraph` enables only the contracts surface — no tree-sitter, no `links` conflict; `monitor-tui` powers `trusty-search monitor tui`; `embedder-client` provides EmbedderClient/RemoteEmbedderClient/UdsEmbedderClient (issue #164); `bm25` is the canonical lexical index (issue #156); `migrations` provides the shared schema-migration kernel (issue #179); `cli-help` provides the shared declarative CLI help + "did you mean?" suggester (issue #216); `update-check` provides throttled crates.io update notification; `bug-capture` enables Phase 1 ERROR event capture to JSONL + in-memory ring (bug-reporting #478)
 
-# ── Bundled sidecar ─────────────────────────────────────────────────────────
+# ── Bundled sidecars ─────────────────────────────────────────────────────────
 # Why: the trusty-embedderd binary is a required runtime dependency (issue
 # #187 Phase 2). Declaring trusty-embedderd as a Cargo dependency causes
 # `cargo install trusty-search` to also compile and install the
@@ -58,6 +66,10 @@ trusty-common = { version = "0.14.1", default-features = false, features = ["axu
 # two binaries. The shim at src/bin/trusty-embedderd.rs calls
 # `trusty_embedderd::run()` so all daemon logic stays in the library crate.
 trusty-embedderd = { workspace = true }
+# Why: bundle trusty-console (Single-Install Convention, epic #943). Declaring
+# it as a dep causes `cargo install trusty-search` to also install trusty-console.
+# All dashboard logic stays in the trusty-console library crate.
+trusty-console = { workspace = true }
 
 # ── Async runtime / HTTP ───────────────────────────────────────────────────
 tokio = { version = "1", features = ["full"] }
